Improve the post-install script
